bitmap2svg/pyproject.toml
[project]
name = "bitmap2svg"
version = "0.1.0"
description = "Hybrid Potrace + geometry + swarm vectoriser for flat-colour logos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.10",
  "Pillow>=10.3",
  "svgwrite>=1.4",
  "shapely>=2.0",
  "pyclipper>=1.3",
  "rdp>=0.8",
  "pydantic>=2.5",
  "typer>=0.12",
  "rich>=13.7",
  "potrace-wrapped; platform_system != 'Windows'",
  "watchdog>=4.0",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "pytesseract>=0.3; extra == 'ocr'",
  "rapidfuzz>=3.9; extra == 'ocr'",
  "torch; extra == 'diff'",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3", "rapidfuzz>=3.9"]
diff = ["torch"]

[project.scripts]
bitmap2svg = "bitmap2svg.cli:app"
bitmap2svg-watch = "bitmap2svg.workers.watcher:main"
bitmap2svg-serve = "bitmap2svg.service:main"

[tool.pytest.ini_options]
testpaths = ["tests"]